Allow stream sizes large enough (#20)
- for both
    - decimal and
    - binary

- issues: https://github.com/TomTheBear/likwid-bench/issues/19 and https://github.com/TomTheBear/likwid-bench/issues/18

+++ include/helper.h
@@ -3,6 +3,6 @@
 #define HELPER_H
 
 double convertToSeconds(const_bstring input);
-long long convertToBytes(const_bstring input);
+size_t convertToBytes(const_bstring input);
 
 #endif /* HELPER_H */
